+++ src/DashboardUI/src/pages/application/review/form/ApplicationReviewFormLayout.tsx
@@ -1,0 +1,15 @@
+import { useOutletContext } from 'react-router-dom';
+import { ApplicationReviewForm, ApplicationReviewFormProps } from './ApplicationReviewForm';
+
+function ApplicationReviewFormLayout() {
+  const { isApplicationLoading, isFundingOrganizationLoading } = useOutletContext<ApplicationReviewFormProps>();
+
+  return (
+    <ApplicationReviewForm
+      isApplicationLoading={isApplicationLoading}
+      isFundingOrganizationLoading={isFundingOrganizationLoading}
+    />
+  );
+}
+
+export default ApplicationReviewFormLayout;
